+++ Frankenpets/Assets/Prefabs/Character/Rigged Characters.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cc9ba53b9355e4b80ad2e928374a4860
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
