+++ Frankenpets/Assets/Prefabs/Character/Rigged Characters/Pet.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 12ced9cb9e79d4298a5f9c8f77c266dc
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
