+++ Frankenpets/Assets/Prefabs/Character/Rigged Characters/catFront.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 94cfde32e1158443b87cf364882672e4
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
