+++ Frankenpets/Assets/Prefabs/Character/Rigged Characters/dogBack.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 69f6a7ec83e324a23bec8d4412cca854
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
