+++ Frankenpets/Assets/Scripts/RiggingMovement.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: c780e751b01a541208b57ba4e2318703
