--- Frankenpets/Assets/Scripts/TestMovement.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: dae41a2fccb0d4fb882b08192def03f2
